Refactoring of Application startup.
Fixed invalid imports.

Added config file settings (pid-file and log-file location).

Moved around some of the startup code in OANApplication class. The
start_server function now creates all objects instead of the
constructor. Makes it possible for the daemon to run that function, and
it's no problem that the Server class creation starts twisted, and it's
own event loop.